+++ types/eslint.d.ts
@@ -1,0 +1,4 @@
+declare module '@typescript-eslint/parser'
+declare module 'eslint-plugin-consistent-default-export-name'
+declare module 'eslint-plugin-node-import'
+declare module 'eslint-plugin-prettier/recommended'